ISharesHolder tests
